編集&削除機能の実装

+++ backend/resources/views/contacts/show.blade.php
@@ -1,8 +1,8 @@
 @extends('layout')
 
 @section('content')
-  <table class="table table-responsive">
-    <tbody>
+  <table class="table">
+    <tbody class="table-responsive">
       <tr class="table-bordered">
         <th class="contact-new_label pr-5">日付</th>
         <td>{{$contact_id->updated_at}}({{$week}})</td>
@@ -21,19 +21,26 @@
         <textarea disabled="disabled" name="mytextarea"  cols="70" rows="10" class="contact-show_body bg-white">{{$contact_id->body}}</textarea>
         </td>
       </tr>
+
+
+      <tr class="border-0">
+        <th class="pr-5"></th>
+        <td class="text-right">
+          @if($user->id == $contact_id->user_id)
+            {{ Form::open(['method' => 'delete', 'route' => ['contact.destroy', $contact_id->id]]) }}
+              {{ Form::submit('[× 削除する]', ['class' => 'btn clear-decoration','style' =>'color: #ef7709;']) }}
+            {{ Form::close() }}
+          @endif
+          <div class="text-center">
+            <a href={{ route('contact.index') }}>
+              <button type="button" class="btn btn-secondary pr-4 pl-4">戻る</button>
+            </a>
+            <a href={{ route('contact.edit',['id'=>$contact_id->id]) }}>
+              <button type="button" class="btn text-white pr-4 pl-4" style ='background: #ef7709;'>編集</button>
+            </a>
+          </div>
+        </td>
+      </tr>
     </tbody>
   </table>
-  {{ Form::open(['method' => 'delete', 'route' => ['contact.destroy', $contact_id->id]]) }}
-  {{ Form::submit('削除') }}
-    <p class="aa text-right pr-5" style ='color: #ef7709;'>[× 削除する]</p>
-  {{ Form::close() }}
-    
-  <div class="text-center">
-    <a href={{ route('contact.index') }}>
-      <button type="button" class="btn btn-secondary pr-4 pl-4">戻る</button>
-    </a>
-    <a href={{ route('contact.edit',['id'=>$contact_id->id]) }}>
-      <button type="button" class="btn text-white pr-4 pl-4" style ='background: #ef7709;'>編集</button>
-    </a>
-  </div>
 @endsection
